requirements.txt
pyttsx3==2.90
SpeechRecognition==3.10.0
openai==1.3.0
python-dotenv==1.0.0
pyaudio==0.2.11